Ensure toggle anchors work with the spacebar
Some assistive technologies, like screen readers, suggest to press 'space'
when interacting with a link with a role of 'button'.
We need to ensure that we replicate this functionality that exists on a button element.

This function is declared outside the main Toggle class as it is not part of the public API.