Ensure anchor toggles are not draggable
If a user drags their mouse slightly when trying to interact with the toggle
it will trigger the 'drag and drop' functionality.
Regular buttons prevent this and ensure as long as the mouse is still above the
button that the click will register.
This will help users with motor impairments and those less familiar with a trackpad.